Refactored dhanhq class to inherit modular classes each having subset of its features
